Updated part client of unit project and adding the test end to end

+++ client/src/js/services/ProjectService.js
@@ -1,15 +1,16 @@
 angular.module('bhima.services')
 .service('ProjectService', ProjectService);
 
-ProjectService.$inject = [ '$http', 'util' ];
+ProjectService.$inject = [ '$http', '$translate', 'util' ];
 
-function ProjectService($http, util) {
+function ProjectService($http, $translate, util) {
   var service = this;
 
   service.create = create;
   service.read = read;
   service.update = update;
-  service.delete = del;
+  service.del = del;
+  service.readComplete = readComplete;
 
   function create(project) {
     return $http.post('/projects', project)
@@ -23,13 +24,24 @@ function ProjectService($http, util) {
     .then(util.unwrapHttpResponse);
   }
 
+  function readComplete() {
+    var url = '/projects?complete=1';
+
+    return $http.get(url)
+    .then(util.unwrapHttpResponse);
+  }
+
   function update(id, project) {
     return $http.put('/projects/' + id, project)
     .then(util.unwrapHttpResponse);
   }
 
   function del(id) {
+
+    var result = confirm($translate.instant('PROJECT.CONFIRM'));
+    if (result) {
     return $http.delete('/projects/' + id)
     .then(util.unwrapHttpResponse);
+    }    
   }
 }
